refactore: 응답자 답 선택 화면 서버 컴포넌트로 수정
